+++ src/main/java/ru/improve/potato/controllers/SessionController.java
@@ -1,34 +1,31 @@
 package ru.improve.potato.controllers;
 
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import ru.improve.potato.services.security.AuthenticationService;
+import ru.improve.potato.dto.session.LoginRequest;
+import ru.improve.potato.dto.session.LoginResponse;
 import ru.improve.potato.validators.auth.AuthValidator;
 
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/auth")
 @Slf4j
-public class SessionAuthController {
+public class SessionController {
 
-    private final AuthenticationService authenticationService;
     private final AuthValidator authValidator;
 
-    /*@PostMapping("/signup")
-    public SignUpResponse signUp(@RequestBody @Valid SingUpRequest singUpRequest,
-                                  BindingResult bindingResult) {
-
-        authValidator.validate(singUpRequest, bindingResult);
-        return authenticationService.signUp(singUpRequest);
-    }
-
-    @PostMapping("/login")
+    @PostMapping()
     public LoginResponse login(@RequestBody @Valid LoginRequest loginRequest,
                                BindingResult bindingResult) {
 
         authValidator.validate(loginRequest, bindingResult);
-        return authenticationService.login(loginRequest);
-    }*/
+
+
+    }
 }
