+++ src/main/java/ru/improve/potato/dto/session/LoginRequest.java
@@ -1,22 +1,16 @@
-package ru.improve.potato.dto.auth;
+package ru.improve.potato.dto.session;
 
 import jakarta.validation.constraints.NotEmpty;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
+import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.extern.jackson.Jacksonized;
 
-@Getter
-@Setter
+@Data
 @AllArgsConstructor
 @NoArgsConstructor
-@Builder
-@Jacksonized
 public class LoginRequest {
 
-    @NotEmpty(message = "telephoneNumber cannot be empty")
+    @NotEmpty(message = "email cannot be empty")
     private String email;
 
     @NotEmpty(message = "password cannot be empty")
